Move sandbox invocation execution tests
